fix(react): split ET reload entry by thread
